minor avatar change,moved to components After Login folder
